implement several Cairo0 hint
is_nn hint uses the assert_felt_le beneath it, but
it was implemented beforehand.

This PR has no tests included since #204 is not solved yet.
I used a couple of Cairo0 scripts to test this functionality
with a set of different arguments to cover both hints
that are a part of `is_nn` function.
(One of them handles negatives while another is for the non-negatives.)

Refs #164